Implement provisioning profiles for CI pipeline
Due to the lack of foreseen solutions, the official repository has
been failing the CI pipeline for a lengthy long period. It raises the
risk that CI will not be able to identify integration issues as soon
as possible. With the help of the deploy key, this patch pulls a unique
"provisioning profile" from a private repository. As a result, it is
allowed to examine the official repository without potentially
disclosing the solutions in forked ones.