Support multiple queues and switching command
Overview: This commit introduces "multiple queue" and provides associated switching commands: 'prev' and 'next'.

We first define a new type 'queue_chain_t', which refers to describe a doubly linked list containing all of queues.
Here is the definition:

typedef struct {
    struct list_head head;
    int size;
} queue_chain_t;

'head' points to the head of a doubly linked list, which chains all of queues.
'size' refers to the number of queues are created.

This commit also changes the definition of meta-data of each queue. Here is the modified definition:

typedef struct {
    struct list_head *q;
    struct list_head chain;
    int size;
    int id;
} qcontext_t;

'q' points to the head of the queue.
'chain' is used by chaining the heads of queues.
'size' refers to the length of this queue.
'id' refers to the unique identification number, which is an integer. To each queue, `id` remains the same even if the previous queue is freed.

New command 'prev' and 'next' are self-explanatory, they help to switch current queue.
This commit also allow 'show_queue' to print the ID of current queue.

Close #101